+++ feature-main/workspace/src/main/java/com/seugi/workspace/navigation/WaitingJoinNavigation.kt
@@ -1,9 +1,10 @@
-package com.seugi.workspace.feature
+package com.seugi.workspace.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
+import com.seugi.workspace.feature.WaitingJoinScreen
 
 const val WAITING_JOIN = "waitingJoin"
 
@@ -14,6 +15,6 @@ fun NavController.navigateToWaitingJoin(navOptions: NavOptions? = null) = naviga
 
 fun NavGraphBuilder.waitingJoin(popBackStack: () -> Unit) {
     composable(route = WAITING_JOIN) {
-        com.seugi.workspace.feature.WaitingJoinScreen(popBackStack = popBackStack)
+        WaitingJoinScreen(popBackStack = popBackStack)
     }
 }
